メソッドを外だし
